Issue #82: start integrating QNLocalizedString class for all error messages across the entire project
WARNING: starting from this commit the project is temporarily in unbuildable state. It's inevitable because it's not easy to switch the type in all required places + while I'm at it, I'd like to generally improve the error messages throughout the project.

The last stop currently is LocalStorageManager_p.cpp, line around 2170