Issue #54: introduce IKeychainService interface as a hook for keychain service functionality to be used in tests
It is inconvenient for the tests to use the actual keychain - some services would bug user many times while the test is running so the fake one needs to be implemented for the tests.
This commit also starts the migration of SynchronizationManagerPrivate class to the use of this new interface.