Issue #54: start implementing the collection of counters for highest USNs of full data items received by the local end from the remote end before API rate limit triggering in FakeNoteStore
The purpose is to make it possible to figure out which data items were fully synced before the first API rate limit happened. On API rate limit breach the sync logic should figure out which data items it managed to sync successfully and persist the corresponding sync state. The number of update counts within this sync state should match those counted by FakeNoteStore independently. So, for the testing purposes we count the same thing in two places independently and then compare the results