Add conditionally asynchronous fetching of the list of low and high usns of all stored sync chunks
Such list presumanly can be fetched in the background before it is really needed and then would be available
in memory for very low cost. In some cases it might be not worth the time to read and deserialize the stored
sync chunks themselves only to figure out they don't cover the needed usns range so having the range readily
available would help to avoid that cost.