Add unit test checking cancellation of sync chunks downloading
It turned out that QPromise::addResult has no effect after QFuture::cancel has been called. It is suboptimal
to say at least as it leaves no chance to save the already downloaded sync chunks in this particular case.
It also means cancellation for notes and resources processors also needs to be rethought