Create in-memory databases with different names for each SQLite local storage unit test instance
Hopefully it should prevent problems arising with random test failures when running on a machine
experiencing high CPU usage. Presumably in such cases when a new test starts, the connection from
the previous test is still alive and so is the in-memory database which it represents. If each test
uses its own in-memory database, this problem should not have a way to exist